HBASE-676 Bytes.getInt returns a long
git-svn-id: https://svn.apache.org/repos/asf/hadoop/hbase/trunk@666260 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hadoop/hbase/util/Bytes.java
@@ -133,9 +133,9 @@ public class Bytes {
    * @param bytes
    * @return the long value
    */
-  public static long toInt(byte[] bytes) {
+  public static int toInt(byte[] bytes) {
     if (bytes == null || bytes.length == 0) {
-      return -1L;
+      return -1;
     }
     return ByteBuffer.wrap(bytes).getInt();
   }
